Started work on the code to create re-sampled Ca-response datasets, tweaking various parameters of the population responses to see how odour separability is affected by these parameters.
cell_trace_dumper.m accumulates real, dF/F trace data and dumps to disk in a certain format, along with info about significant responses and stimulus frames.

cell_pop_simulator.m is the beginning of the mother program that will call various functions to run the entire analysis. Currently, it loads in raw data dumped by cell_trace_dumper and calls resample_cell_pop.m

resample_cell_pop.m is the beginning of a function to take raw data and some parameters that describe a population's responses and re-sample the raw data to create an artificial set of responses. The parameters it takes as inputs are n_cells, n_odors, sparseness_mean, sparseness_sd (this is the frac of cells responsive to a given odour). Also need to come up with a way to specify how likely a cell is to respond to multiple odours (skew away from random in both directions).