cell_pop_simulator: minor changes
resample_cell_pop: finished work on program that first creates a matrix of significant response cell-odor pairs and then re-samples raw data according to this significant response matrix. Responses to all odor durations are treated as a single unit. Input parameters for creating the sig resp matrix are n_cells, n_odors, mean_sparseness, sd_sparseness and cooperativity. A vector of length n_odors of randomly chosen sparseness values from a normal dist with specified mean and SD specifies the numbers of cells responding to each odor.
Cooperativity determines which cells are chosen to be responsive to the multiple odors. For positive cooperativities, a matrix is initialised with all odor responses being assigned to the same set of cells, ie many cells respond to 0 odors for sparsenesses < 1. If cooperativity is 1, this is the matrix used to resample raw data. If less than 1, the appropriate fraction of responding cell-odor pairs are reassigned to random cell-odor pairs. If cooperativity is 0, all cell-odor pair responses are chosen randomly. For negative cooperativities, the matrix is initialised with all odor responses distributed systematically across all cells, ie, all cells respond to roughly the same number of odors.