updated KC param fitting script to handle nans in raw data gracefully
Most raw trace datasets have trials that are padded with nans to account for different numbers of frames in different trials of a given dataset. Now the fitting script identifies this set of padding nans and removes them from the trials before fitting. Trials that are more than 70% nans are excluded from fitting - these are probably bad trials excluded due to z-movement.